+++ include/runtime/oxygen_runtime.h
@@ -2,7 +2,6 @@
 #define OXYGEN_RT_H
 
 
-#include <cstddef>
 #include <stdlib.h>
 #include <string.h>
 #include <stdint.h>
@@ -11,14 +10,14 @@
 
 #define OXYGEN_VERSION_MAJOR 0
 #define OXYGEN_VERSION_MINOR 1
-#define OXYGEN_VERSION_PATCH 0\
+#define OXYGEN_VERSION_PATCH 0
 
 #define OXYGEN_VERSION "0.1.0"
 
 // HASH64
 
-unsigned long long hash64(long n) {
-    unsigned long long hash= clock() ^ 13 + n;
+unsigned long long oxygen_hash64(long n) {
+    unsigned long long hash= (n<<11) ^ 13 + n;
     return hash;
 }
 
@@ -28,14 +27,16 @@ typedef struct string {
     unsigned int capacity;
 } string;
 
-void new_string(string* s, const char* str ) {
+
+
+void oxygen_new_string(string* s, const char* str ) {
     s->size = strlen(str);
     s->capacity = s->size * 2+1;
     s->data = (char*)malloc(s->capacity+1);
     strcpy(s->data, str);
 }
 
-void string_concat(string* s, const char* str) {
+void oxygen_string_concat(string* s, const char* str) {
     unsigned int len = strlen(str);
     if(s->capacity<=(s->size+len+1)) {
         s->data = (char*)realloc(s->data, s->capacity+len+1);
@@ -45,14 +46,13 @@ void string_concat(string* s, const char* str) {
     s->size += len;
 }
 
-void string_reserve(string* s, unsigned int size) {
+void oxygen_string_reserve(string* s, unsigned int size) {
     s->data = (char*)realloc(s->data, s->capacity+size);
     s->capacity += size;
 }
 
-void string_deleter(string* s) {
+void oxygen_string_destroy(string* s) {
     free(s->data);
-    free(s);
 }
 
 // ptr_rt
@@ -75,5 +75,9 @@ void* oxygen_move(void **a) {
         unsigned long long capacity; \
     } vec;
 
+
+// Other
+
+#define REF(x) (&(x))
 
 #endif