+++ src/runtime/test.c
@@ -3,13 +3,10 @@
 #include <stdio.h>
 
 int main() {
-    string* str = (string*)malloc(sizeof(string));
-    new_string(str, "Hi!");
-    string_concat(str, " Im man!");
-    unique_ptr* p = make_unique(str);
-    printf("%s\n", p->val.str->data);
-    free(str->data);
-    free(str);
-    destroy_unique(p);
+    string str;
+    new_string(&str, "Hi!");
+    string_concat(&str, " Im man!");
+    printf("%s\n", str.data);
+    free(str.data);
     return 0;
 }