+++ src/table.cpp
@@ -31,6 +31,7 @@ symbol search(const std::string &name) {
 
 
 
+
 symbol* searchptr(const std::string &name) {
     for(auto &scope : std::ranges::reverse_view(table)) {
         if(scope.contains(name)) {
@@ -56,7 +57,7 @@ token_type search_type_scope(const std::string &name, unsigned int lvl) {
 
 void insert(const std::string &name,token_type type, token_value val, bool is_const, u64 size, bool is_array, bool comptime, bool is_vector, bool isptr, const std::string &modname, bool ismov, bool isref) {
     if(table.empty()) [[unlikely]] return;
-    table[table.size()-1].insert_or_assign(name, symbol{type, std::move(val), is_const, size, is_array, comptime,name, is_vector, isptr, modname, ismov});
+    table[table.size()-1].insert_or_assign(name, symbol{type, std::move(val), is_const, size, is_array, comptime,name, is_vector, isptr, modname, ismov, !isref,  isref});
 }
 
 void insert_top(const std::string &name,token_type type,token_value val, bool is_const, u64 size,bool is_array, bool comptime, bool is_vector, bool isptr, const std::string modname) {