+++ changelog.md
@@ -26,11 +26,13 @@
 * build time
 
 
-### Flame 0.10.0(13.07.26):
+### Flame 0.10.0(17.07.26):
 
 * mutable and not mutable references(**function arguments only**)
 * new style of warnings
 * now only one mutable refernce can be defined for var
+* switching to C backend
+* -t flag for only transpilation
 
 * **BUGFIXES**:
     
@@ -40,3 +42,8 @@
     * bug was making that argument type and provided variable type even if mismatch it was working
     * string tokens didnt have .str_value
     * accessing string with compile time index 
+
+### Shiver 1.1
+
+* Added Flameflags parameter for build
+* minor changes